Testing (#1)
* Start adding acceptance tests

* Add test for permission datasource

* Add test for role data source

* Add test for system channel data source

* Add test for catergory channel resource

* Add test for text channel resource

* Add test for invite resource

* Simplfy checking for string

* Add test for channel permission resource

* Add test for webhook resource

* Add test for message resource

* Add test for role resource

* Finish unit tests

+++ discord/resource_discord_role_everyone.go
@@ -67,7 +67,7 @@ func resourceRoleEveryoneUpdate(ctx context.Context, d *schema.ResourceData, m i
 
 	serverId := d.Get("server_id").(string)
 	d.SetId(serverId)
-	newPermission := d.Get("permissions").(int64)
+	newPermission := int64(d.Get("permissions").(int))
 
 	if role, err := client.GuildRoleEdit(serverId, serverId, &discordgo.RoleParams{
 		Permissions: &newPermission,
